requirements.txt
astroid==2.2.5
attrs==19.3.0
Automat==0.8.0
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
docopt==0.6.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.2
lxml==4.4.1
mccabe==0.6.1
numpy==1.17.3
pandas==0.25.2
parsel==1.5.2
pipreqs==0.4.9
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pylint==2.3.1
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2019.3
queuelib==1.5.0
requests==2.22.0
Scrapy==1.7.4
scrapy-crawlera==1.6.0
scrapy_redis==0.6.8
service-identity==18.1.0
six==1.12.0
Twisted==19.7.0
typed-ast==1.4.0
urllib3==1.25.6
w3lib==1.21.0
weibo==0.2.2
wrapt==1.11.2
yarg==0.1.9
zope.interface==4.6.0
pymysql==0.9.3
